Remove profile view, disable thread pool for profiling
